Refactored all the code.

+++ src/ShvetsGroup/Command/CleanupCommand.php
@@ -3,25 +3,37 @@
 namespace ShvetsGroup\Command;
 
 use Symfony\Component\Console as Console;
+use ShvetsGroup\Service\JobsManager;
+use ShvetsGroup\Service\Proxy;
 
 class CleanupCommand extends Console\Command\Command
 {
 
     /**
-     * @var \ShvetsGroup\Service\JobsManager
+     * @var JobsManager
      */
     private $jobsManager;
 
     /**
-     * @param \ShvetsGroup\Service\JobsManager $jobsManager
+     * @var Proxy
      */
-    public function __construct($jobsManager)
+    private $proxy;
+
+    /**
+     * @param JobsManager $jobsManager
+     * @param Proxy $proxy
+     */
+    public function __construct($jobsManager, $proxy)
     {
         parent::__construct('cleanup');
 
         $this->setDescription('Cleanup jobs pool.');
+        $this->addOption('all', 'a', Console\Input\InputOption::VALUE_NONE, 'Kill all proxies and jobs.');
+        $this->addOption('proxy', 'p', Console\Input\InputOption::VALUE_NONE, 'Kill all proxies.');
+        $this->addOption('jobs', 'j', Console\Input\InputOption::VALUE_NONE, 'Flush all jobs.');
 
         $this->jobsManager = $jobsManager;
+        $this->proxy = $proxy;
     }
 
     /**
@@ -35,7 +47,12 @@ class CleanupCommand extends Console\Command\Command
      */
     protected function execute(Console\Input\InputInterface $input, Console\Output\OutputInterface $output)
     {
-        $this->jobsManager->deleteAll();
+        if ($input->getOption('jobs') || $input->getOption('all')) {
+            $this->jobsManager->deleteAll();
+        }
+        if ($input->getOption('proxy') || $input->getOption('all')) {
+            $this->proxy->killAll();
+        }
 
         return true;
     }
